pyproject.toml
[project]
name = "hypertube-tests"
version = "0.1.0"
description = "Tests for Hypertube XP Cards"
requires-python = ">=3.12"
dependencies = [
    "pytest>=8.3",
    "requests>=2.32",
    "psycopg2-binary>=2.9",
    "pyjwt>=2.9",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"

[dependency-groups]
dev = [
    "pytest>=8.3",
    "requests>=2.32",
    "psycopg2-binary>=2.9",
    "pyjwt>=2.9",
]
